move "current version" body classes and advisory note to module.version
