+++ lib/core/init.js
@@ -1,5 +1,3 @@
-import { $ } from '../vendor';
-import * as Metadata from './metadata';
 import * as Modules from './modules';
 import { _loadModuleOptions } from './options/options';
 import { _loadModulePrefs } from './modules/modules';
@@ -12,7 +10,6 @@ import { // eslint-disable-line import/order
 	map,
 	nonNull,
 	now,
-	range,
 	waitForChild,
 	waitForEvent,
 } from '../utils';
@@ -153,43 +150,10 @@ function _addModuleBodyClasses() {
 	}
 }
 
-sourceLoaded.then(addVersionClasses);
-
 documentReady.then(() => BodyClasses.add());
 bodyStart.then(() => BodyClasses.add());
 
 Promise.all([bodyReady, go]).then(initObservers);
-
-bodyReady.then(reportVersion);
-
-bodyReady.then(homePage);
-
-function addVersionClasses() {
-	BodyClasses.add('res', 'res-v430');
-	const versionComponents = Metadata.version.split('.');
-	for (const i of range(0, versionComponents.length)) {
-		BodyClasses.add(`res-v${versionComponents.slice(0, i + 1).join('-')}`);
-	}
-}
-
-function reportVersion() {
-	// report the version of RES to reddit's advisory checker.
-	$('<div>', {
-		id: 'RESConsoleVersion',
-		style: 'display: none;',
-		text: Metadata.version,
-	}).appendTo(document.body);
-}
-
-function homePage() {
-	if (location.href.includes('reddit.honestbleeps.com/download') || location.href.includes('redditenhancementsuite.com/download')) {
-		for (const link of document.body.querySelectorAll('.install')) {
-			link.classList.add('update');
-			link.classList.add('res4'); // if update but not RES 4, then FF users == greasemonkey...
-			link.classList.remove('install');
-		}
-	}
-}
 
 function setupProfiling() {
 	const end = {};
